requirements_all.txt
absl-py==1.4.0
astunparse==1.6.3
attrs==23.1.0
basemap==1.3.7
basemap-data==1.3.2
branca==0.6.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.3
Fiona==1.9.4.post1
flatbuffers==23.5.26
folium==0.14.0
fonttools==4.40.0
gast==0.5.4
geographiclib==2.0
geopandas==0.13.2
geopy==2.4.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.2
h11==0.14.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keras-tcn==3.4.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
networkx==3.1
numpy==1.20.0
oauthlib==3.2.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==1.3.0
patsy==0.5.3
Pillow==9.5.0
protobuf==3.20.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.0
pyproj==3.5.0
pyshp==2.3.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.3.3
seaborn==0.12.2
selenium==4.11.2
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
statsmodels==0.12.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.20.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
torch==1.10.1
trio==0.22.2
trio-websocket==0.10.3
typeguard==2.13.3
typing-extensions==4.6.3
tqdm==4.66.1
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0